requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
streamlit
kaggle
python-dotenv

